Rename cron to daemon:
* Crons aren't really crons, they are just loop+delay
* DueTimerConsumer is not a cron, it's an event driven loop
