+++ server/daemons/imminent-child-run-wait-timed-out-runs.ts
@@ -1,9 +1,9 @@
 import type { NonEmptyArray } from "@aikirun/lib/array";
 import { chunkLazy, isNonEmptyArray } from "@aikirun/lib/array";
-import { streamChunks } from "@aikirun/lib/async";
-import type { WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
+import type { WorkflowRunState, WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
 import type { WorkflowRunMeta } from "server/infra/db/pg/repository/workflow-run";
 import type {
+	ChildWorkflowRunWaitQueueRowInsert,
 	Repositories,
 	StateTransitionRowInsert,
 	WorkflowRow,
@@ -12,71 +12,55 @@ import type {
 import type { WorkflowRunPublisher } from "server/infra/messaging/redis-publisher";
 import type { TimerEntry, TimerSortedSet } from "server/infra/messaging/redis-timer-sorted-set";
 import { runConcurrently } from "server/lib/concurrency";
-import type { CronContext } from "server/middleware/context";
+import type { DaemonContext } from "server/middleware/context";
 import { ulid } from "ulidx";
 
-import { createTimerStreamCursorAdvancer } from "./lib/timer-stream";
+import { streamTimers } from "./lib/timer-stream";
 import { publishRuns } from "./publish-ready-runs";
 
 type Repos = Pick<
 	Repositories,
-	"task" | "workflowRun" | "stateTransition" | "workflow" | "workflowRunOutbox" | "transaction"
+	"workflowRun" | "stateTransition" | "childWorkflowRunWaitQueue" | "workflow" | "workflowRunOutbox" | "transaction"
 >;
 
-export interface ProcessImminentRetryableTaskRunsDeps {
+export interface ProcessImminentChildRunWaitTimedOutRunsDeps {
 	repos: Repos;
 	workflowRunPublisher?: WorkflowRunPublisher;
 	timerSortedSet?: TimerSortedSet;
 }
 
-const advanceTaskCursor = createTimerStreamCursorAdvancer<{ workflowRunId: string; dueAt: Date }>({
-	getDueAt: (entry) => entry.dueAt,
-	getId: (entry) => entry.workflowRunId,
-});
-
-export async function processImminentRetryableTaskRuns(
-	context: CronContext,
-	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentRetryableTaskRunsDeps,
+export async function processImminentChildRunWaitTimedOutRuns(
+	context: DaemonContext,
+	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentChildRunWaitTimedOutRunsDeps,
 	options?: { limit?: number; imminenceThresholdMs?: number }
 ) {
 	const { limit = 1_000, imminenceThresholdMs = 5_000 } = options ?? {};
 
 	const dueBefore = new Date(Date.now() + imminenceThresholdMs);
 
-	let now = Date.now();
-	for await (const { whenTrue: tasksDueNow, whenFalse: tasksDueSoon } of streamChunks(
-		(cursor) => repos.task.listRetryableTaskWorkflowRuns(context, dueBefore, limit, cursor),
-		{
-			advanceCursor: advanceTaskCursor,
-			until: (chunk) => chunk.length < limit,
-			partition: (task: { dueAt: Date }) => task.dueAt.getTime() <= now,
-		}
+	for await (const { dueNow: runsDueNow, dueSoon: runsDueSoon } of streamTimers(
+		(cursor) => repos.workflowRun.listChildRunWaitTimedOutRuns(context, dueBefore, limit, cursor),
+		(chunk) => chunk.length < limit
 	)) {
-		if (isNonEmptyArray(tasksDueNow)) {
-			const runIds = tasksDueNow.map((task) => task.workflowRunId) as NonEmptyArray<string>;
-			const runs = await repos.workflowRun.listByIdsAndStatus(context, runIds, "running");
-			if (isNonEmptyArray(runs)) {
-				await queueRetryableTaskRuns(context, repos, workflowRunPublisher, runs);
-			}
+		if (isNonEmptyArray(runsDueNow)) {
+			await queueChildRunWaitTimedOutRuns(context, repos, workflowRunPublisher, runsDueNow);
 		}
 
-		if (timerSortedSet && isNonEmptyArray(tasksDueSoon)) {
-			const timers: TimerEntry[] = tasksDueSoon.map((task) => ({
-				type: "task_retry",
-				id: task.workflowRunId,
-				dueAt: task.dueAt.getTime(),
+		if (timerSortedSet && isNonEmptyArray(runsDueSoon)) {
+			const timers: TimerEntry[] = runsDueSoon.map((run) => ({
+				type: "child_wait_timeout",
+				id: run.id,
+				dueAt: run.dueAt.getTime(),
 			}));
 			if (isNonEmptyArray(timers)) {
 				await timerSortedSet.add(timers);
 			}
 		}
-
-		now = Date.now();
 	}
 }
 
-export async function queueRetryableTaskRuns(
-	context: CronContext,
+export async function queueChildRunWaitTimedOutRuns(
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
@@ -84,16 +68,28 @@ export async function queueRetryableTaskRuns(
 ) {
 	const { chunkSize = runs.length } = options ?? {};
 
-	const workflowIds = Array.from(new Set(runs.map((run) => run.workflowId)));
-	if (!isNonEmptyArray(workflowIds)) {
+	const stateTransitionIds: string[] = [];
+	const workflowIdSet = new Set<string>();
+	for (const run of runs) {
+		stateTransitionIds.push(run.latestStateTransitionId);
+		workflowIdSet.add(run.workflowId);
+	}
+	const workflowIds = Array.from(workflowIdSet);
+
+	if (!isNonEmptyArray(stateTransitionIds) || !isNonEmptyArray(workflowIds)) {
 		return;
 	}
-	const workflows = await repos.workflow.getByIdsGlobal(context, workflowIds);
+
+	const [stateTransitions, workflows] = await Promise.all([
+		repos.stateTransition.getByIds(stateTransitionIds),
+		repos.workflow.getByIdsGlobal(context, workflowIds),
+	]);
+	const stateTransitionsById = new Map(stateTransitions.map((transition) => [transition.id, transition]));
 	const workflowsById = new Map(workflows.map((workflow) => [workflow.id, workflow]));
 
 	await runConcurrently(context, chunkLazy(runs, chunkSize), async (chunk, spanCtx) => {
 		try {
-			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, workflowsById);
+			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, stateTransitionsById, workflowsById);
 		} catch (error) {
 			spanCtx.logger.warn({ error, chunkSize: chunk.length }, "Failed to process chunk, will retry next tick");
 		}
@@ -101,12 +97,16 @@ export async function queueRetryableTaskRuns(
 }
 
 async function processChunk(
-	context: CronContext,
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
+	stateTransitionsById: Map<string, { id: string; state: unknown }>,
 	workflowsById: Map<string, WorkflowRow>
 ): Promise<void> {
+	const timedOutAt = new Date();
+
+	const childRunWaitEntries: ChildWorkflowRunWaitQueueRowInsert[] = [];
 	const stateTransitionEntries: StateTransitionRowInsert[] = [];
 	const workflowRunUpdates: Array<{ filter: { id: string; revision: number }; update: { stateTransitionId: string } }> =
 		[];
@@ -118,15 +118,33 @@ async function processChunk(
 			continue;
 		}
 
+		const transition = stateTransitionsById.get(run.latestStateTransitionId);
+		if (!transition) {
+			continue;
+		}
+		const fromState = transition.state as WorkflowRunState;
+		if (fromState.status !== "awaiting_child_workflow") {
+			continue;
+		}
+
+		childRunWaitEntries.push({
+			id: ulid(),
+			parentWorkflowRunId: run.id,
+			childWorkflowRunId: fromState.childWorkflowRunId,
+			childWorkflowRunStatus: fromState.childWorkflowRunStatus,
+			status: "timeout",
+			timedOutAt,
+		});
+
 		const stateTransitionId = ulid();
-		const state: WorkflowRunStateQueued = { status: "queued", reason: "task_retry" };
+		const toState: WorkflowRunStateQueued = { status: "queued", reason: "child_workflow" };
 		stateTransitionEntries.push({
 			id: stateTransitionId,
 			workflowRunId: run.id,
 			type: "workflow_run",
 			status: "queued",
 			attempt: run.attempts,
-			state,
+			state: toState,
 		});
 		workflowRunUpdates.push({
 			filter: {
@@ -137,7 +155,6 @@ async function processChunk(
 				stateTransitionId,
 			},
 		});
-
 		outboxEntries.push({
 			id: ulid(),
 			namespaceId: run.namespaceId,
@@ -149,13 +166,21 @@ async function processChunk(
 		});
 	}
 
-	if (!isNonEmptyArray(stateTransitionEntries) || !isNonEmptyArray(workflowRunUpdates)) {
+	if (
+		!isNonEmptyArray(childRunWaitEntries) ||
+		!isNonEmptyArray(stateTransitionEntries) ||
+		!isNonEmptyArray(workflowRunUpdates)
+	) {
 		return;
 	}
 
 	const insertedOutboxEntries: WorkflowRunOutboxRowInsert[] = await repos.transaction(async (txRepos) => {
+		await txRepos.childWorkflowRunWaitQueue.insert(childRunWaitEntries);
 		await txRepos.stateTransition.appendBatch(stateTransitionEntries);
-		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued("running", workflowRunUpdates);
+		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued(
+			"awaiting_child_workflow",
+			workflowRunUpdates
+		);
 		const transitionedRunIdsSet = new Set(transitionedRunIds);
 		const outboxEntriesToInsert = outboxEntries.filter((entry) => transitionedRunIdsSet.has(entry.workflowRunId));
 		if (!isNonEmptyArray(outboxEntriesToInsert)) {
