+++ server/daemons/imminent-retryable-runs.ts
@@ -1,9 +1,8 @@
 import type { NonEmptyArray } from "@aikirun/lib/array";
 import { chunkLazy, isNonEmptyArray } from "@aikirun/lib/array";
-import type { WorkflowRunState, WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
+import type { WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
 import type { WorkflowRunMeta } from "server/infra/db/pg/repository/workflow-run";
 import type {
-	EventWaitQueueRowInsert,
 	Repositories,
 	StateTransitionRowInsert,
 	WorkflowRow,
@@ -12,7 +11,7 @@ import type {
 import type { WorkflowRunPublisher } from "server/infra/messaging/redis-publisher";
 import type { TimerEntry, TimerSortedSet } from "server/infra/messaging/redis-timer-sorted-set";
 import { runConcurrently } from "server/lib/concurrency";
-import type { CronContext } from "server/middleware/context";
+import type { DaemonContext } from "server/middleware/context";
 import { ulid } from "ulidx";
 
 import { streamTimers } from "./lib/timer-stream";
@@ -20,18 +19,18 @@ import { publishRuns } from "./publish-ready-runs";
 
 type Repos = Pick<
 	Repositories,
-	"workflowRun" | "stateTransition" | "eventWaitQueue" | "workflow" | "workflowRunOutbox" | "transaction"
+	"workflowRun" | "stateTransition" | "task" | "workflow" | "workflowRunOutbox" | "transaction"
 >;
 
-export interface ProcessImminentEventWaitTimedOutRunsDeps {
+export interface ProcessImminentRetryableRunsDeps {
 	repos: Repos;
 	workflowRunPublisher?: WorkflowRunPublisher;
 	timerSortedSet?: TimerSortedSet;
 }
 
-export async function processImminentEventWaitTimedOutRuns(
-	context: CronContext,
-	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentEventWaitTimedOutRunsDeps,
+export async function processImminentRetryableRuns(
+	context: DaemonContext,
+	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentRetryableRunsDeps,
 	options?: { limit?: number; imminenceThresholdMs?: number }
 ) {
 	const { limit = 1_000, imminenceThresholdMs = 5_000 } = options ?? {};
@@ -39,16 +38,16 @@ export async function processImminentEventWaitTimedOutRuns(
 	const dueBefore = new Date(Date.now() + imminenceThresholdMs);
 
 	for await (const { dueNow: runsDueNow, dueSoon: runsDueSoon } of streamTimers(
-		(cursor) => repos.workflowRun.listEventWaitTimedOutRuns(context, dueBefore, limit, cursor),
+		(cursor) => repos.workflowRun.listRetryableRuns(context, dueBefore, limit, cursor),
 		(chunk) => chunk.length < limit
 	)) {
 		if (isNonEmptyArray(runsDueNow)) {
-			await queueEventWaitTimedOutRuns(context, repos, workflowRunPublisher, runsDueNow);
+			await queueRetryableRuns(context, repos, workflowRunPublisher, runsDueNow);
 		}
 
 		if (timerSortedSet && isNonEmptyArray(runsDueSoon)) {
 			const timers: TimerEntry[] = runsDueSoon.map((run) => ({
-				type: "event_wait_timeout",
+				type: "retry",
 				id: run.id,
 				dueAt: run.dueAt.getTime(),
 			}));
@@ -59,8 +58,8 @@ export async function processImminentEventWaitTimedOutRuns(
 	}
 }
 
-export async function queueEventWaitTimedOutRuns(
-	context: CronContext,
+export async function queueRetryableRuns(
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
@@ -68,28 +67,16 @@ export async function queueEventWaitTimedOutRuns(
 ) {
 	const { chunkSize = runs.length } = options ?? {};
 
-	const stateTransitionIds: string[] = [];
-	const workflowIdSet = new Set<string>();
-	for (const run of runs) {
-		stateTransitionIds.push(run.latestStateTransitionId);
-		workflowIdSet.add(run.workflowId);
-	}
-	const workflowIds = Array.from(workflowIdSet);
-
-	if (!isNonEmptyArray(stateTransitionIds) || !isNonEmptyArray(workflowIds)) {
+	const workflowIds = Array.from(new Set(runs.map((run) => run.workflowId)));
+	if (!isNonEmptyArray(workflowIds)) {
 		return;
 	}
-
-	const [stateTransitions, workflows] = await Promise.all([
-		repos.stateTransition.getByIds(stateTransitionIds),
-		repos.workflow.getByIdsGlobal(context, workflowIds),
-	]);
-	const stateTransitionsById = new Map(stateTransitions.map((transition) => [transition.id, transition]));
+	const workflows = await repos.workflow.getByIdsGlobal(context, workflowIds);
 	const workflowsById = new Map(workflows.map((workflow) => [workflow.id, workflow]));
 
 	await runConcurrently(context, chunkLazy(runs, chunkSize), async (chunk, spanCtx) => {
 		try {
-			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, stateTransitionsById, workflowsById);
+			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, workflowsById);
 		} catch (error) {
 			spanCtx.logger.warn({ error, chunkSize: chunk.length }, "Failed to process chunk, will retry next tick");
 		}
@@ -97,16 +84,14 @@ export async function queueEventWaitTimedOutRuns(
 }
 
 async function processChunk(
-	context: CronContext,
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
-	stateTransitionsById: Map<string, { id: string; state: unknown }>,
 	workflowsById: Map<string, WorkflowRow>
 ): Promise<void> {
-	const timedOutAt = new Date();
+	const workflowRunIds = runs.map((run) => run.id);
 
-	const eventWaitEntries: EventWaitQueueRowInsert[] = [];
 	const stateTransitionEntries: StateTransitionRowInsert[] = [];
 	const workflowRunUpdates: Array<{ filter: { id: string; revision: number }; update: { stateTransitionId: string } }> =
 		[];
@@ -118,32 +103,15 @@ async function processChunk(
 			continue;
 		}
 
-		const transition = stateTransitionsById.get(run.latestStateTransitionId);
-		if (!transition) {
-			continue;
-		}
-		const fromState = transition.state as WorkflowRunState;
-		if (fromState.status !== "awaiting_event") {
-			continue;
-		}
-
-		eventWaitEntries.push({
-			id: ulid(),
-			workflowRunId: run.id,
-			name: fromState.eventName,
-			status: "timeout",
-			timedOutAt,
-		});
-
 		const stateTransitionId = ulid();
-		const toState: WorkflowRunStateQueued = { status: "queued", reason: "event" };
+		const state: WorkflowRunStateQueued = { status: "queued", reason: "retry" };
 		stateTransitionEntries.push({
 			id: stateTransitionId,
 			workflowRunId: run.id,
 			type: "workflow_run",
 			status: "queued",
 			attempt: run.attempts,
-			state: toState,
+			state,
 		});
 		workflowRunUpdates.push({
 			filter: {
@@ -167,7 +135,7 @@ async function processChunk(
 	}
 
 	if (
-		!isNonEmptyArray(eventWaitEntries) ||
+		!isNonEmptyArray(workflowRunIds) ||
 		!isNonEmptyArray(stateTransitionEntries) ||
 		!isNonEmptyArray(workflowRunUpdates)
 	) {
@@ -175,9 +143,9 @@ async function processChunk(
 	}
 
 	const insertedOutboxEntries: WorkflowRunOutboxRowInsert[] = await repos.transaction(async (txRepos) => {
-		await txRepos.eventWaitQueue.insert(eventWaitEntries);
+		await txRepos.task.deleteStaleByWorkflowRunIds(workflowRunIds);
 		await txRepos.stateTransition.appendBatch(stateTransitionEntries);
-		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued("awaiting_event", workflowRunUpdates);
+		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued("awaiting_retry", workflowRunUpdates);
 		const transitionedRunIdsSet = new Set(transitionedRunIds);
 		const outboxEntriesToInsert = outboxEntries.filter((entry) => transitionedRunIdsSet.has(entry.workflowRunId));
 		if (!isNonEmptyArray(outboxEntriesToInsert)) {
