+++ server/daemons/imminent-scheduled-runs.ts
@@ -1,6 +1,6 @@
 import type { NonEmptyArray } from "@aikirun/lib/array";
 import { chunkLazy, isNonEmptyArray } from "@aikirun/lib/array";
-import type { WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
+import type { WorkflowRunState, WorkflowRunStateQueued, WorkflowStartOptions } from "@aikirun/types/workflow-run";
 import type { WorkflowRunMeta } from "server/infra/db/pg/repository/workflow-run";
 import type {
 	Repositories,
@@ -11,26 +11,23 @@ import type {
 import type { WorkflowRunPublisher } from "server/infra/messaging/redis-publisher";
 import type { TimerEntry, TimerSortedSet } from "server/infra/messaging/redis-timer-sorted-set";
 import { runConcurrently } from "server/lib/concurrency";
-import type { CronContext } from "server/middleware/context";
+import type { DaemonContext } from "server/middleware/context";
 import { ulid } from "ulidx";
 
 import { streamTimers } from "./lib/timer-stream";
 import { publishRuns } from "./publish-ready-runs";
 
-type Repos = Pick<
-	Repositories,
-	"workflowRun" | "stateTransition" | "task" | "workflow" | "workflowRunOutbox" | "transaction"
->;
+type Repos = Pick<Repositories, "workflowRun" | "workflow" | "stateTransition" | "workflowRunOutbox" | "transaction">;
 
-export interface ProcessImminentRetryableRunsDeps {
+export interface ProcessImminentScheduledRunsDeps {
 	repos: Repos;
 	workflowRunPublisher?: WorkflowRunPublisher;
 	timerSortedSet?: TimerSortedSet;
 }
 
-export async function processImminentRetryableRuns(
-	context: CronContext,
-	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentRetryableRunsDeps,
+export async function processImminentScheduledRuns(
+	context: DaemonContext,
+	{ repos, workflowRunPublisher, timerSortedSet }: ProcessImminentScheduledRunsDeps,
 	options?: { limit?: number; imminenceThresholdMs?: number }
 ) {
 	const { limit = 1_000, imminenceThresholdMs = 5_000 } = options ?? {};
@@ -38,16 +35,16 @@ export async function processImminentRetryableRuns(
 	const dueBefore = new Date(Date.now() + imminenceThresholdMs);
 
 	for await (const { dueNow: runsDueNow, dueSoon: runsDueSoon } of streamTimers(
-		(cursor) => repos.workflowRun.listRetryableRuns(context, dueBefore, limit, cursor),
+		(cursor) => repos.workflowRun.listDueScheduleRuns(context, dueBefore, limit, cursor),
 		(chunk) => chunk.length < limit
 	)) {
 		if (isNonEmptyArray(runsDueNow)) {
-			await queueRetryableRuns(context, repos, workflowRunPublisher, runsDueNow);
+			await queueScheduledRuns(context, repos, workflowRunPublisher, runsDueNow);
 		}
 
 		if (timerSortedSet && isNonEmptyArray(runsDueSoon)) {
 			const timers: TimerEntry[] = runsDueSoon.map((run) => ({
-				type: "retry",
+				type: "scheduled",
 				id: run.id,
 				dueAt: run.dueAt.getTime(),
 			}));
@@ -58,8 +55,8 @@ export async function processImminentRetryableRuns(
 	}
 }
 
-export async function queueRetryableRuns(
-	context: CronContext,
+export async function queueScheduledRuns(
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
@@ -67,16 +64,28 @@ export async function queueRetryableRuns(
 ) {
 	const { chunkSize = runs.length } = options ?? {};
 
-	const workflowIds = Array.from(new Set(runs.map((run) => run.workflowId)));
-	if (!isNonEmptyArray(workflowIds)) {
+	const stateTransitionIds: string[] = [];
+	const workflowIdSet = new Set<string>();
+	for (const run of runs) {
+		stateTransitionIds.push(run.latestStateTransitionId);
+		workflowIdSet.add(run.workflowId);
+	}
+	const workflowIds = Array.from(workflowIdSet);
+
+	if (!isNonEmptyArray(stateTransitionIds) || !isNonEmptyArray(workflowIds)) {
 		return;
 	}
-	const workflows = await repos.workflow.getByIdsGlobal(context, workflowIds);
+
+	const [stateTransitions, workflows] = await Promise.all([
+		repos.stateTransition.getByIds(stateTransitionIds),
+		repos.workflow.getByIdsGlobal(context, workflowIds),
+	]);
+	const stateTransitionsById = new Map(stateTransitions.map((transition) => [transition.id, transition]));
 	const workflowsById = new Map(workflows.map((workflow) => [workflow.id, workflow]));
 
 	await runConcurrently(context, chunkLazy(runs, chunkSize), async (chunk, spanCtx) => {
 		try {
-			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, workflowsById);
+			await processChunk(spanCtx, repos, workflowRunPublisher, chunk, stateTransitionsById, workflowsById);
 		} catch (error) {
 			spanCtx.logger.warn({ error, chunkSize: chunk.length }, "Failed to process chunk, will retry next tick");
 		}
@@ -84,14 +93,13 @@ export async function queueRetryableRuns(
 }
 
 async function processChunk(
-	context: CronContext,
+	context: DaemonContext,
 	repos: Repos,
 	workflowRunPublisher: WorkflowRunPublisher | undefined,
 	runs: NonEmptyArray<WorkflowRunMeta>,
+	stateTransitionsById: Map<string, { id: string; state: unknown }>,
 	workflowsById: Map<string, WorkflowRow>
 ): Promise<void> {
-	const workflowRunIds = runs.map((run) => run.id);
-
 	const stateTransitionEntries: StateTransitionRowInsert[] = [];
 	const workflowRunUpdates: Array<{ filter: { id: string; revision: number }; update: { stateTransitionId: string } }> =
 		[];
@@ -103,15 +111,24 @@ async function processChunk(
 			continue;
 		}
 
+		const transition = stateTransitionsById.get(run.latestStateTransitionId);
+		if (!transition) {
+			continue;
+		}
+		const fromState = transition.state as WorkflowRunState;
+		if (fromState.status !== "scheduled") {
+			continue;
+		}
+
 		const stateTransitionId = ulid();
-		const state: WorkflowRunStateQueued = { status: "queued", reason: "retry" };
+		const toState: WorkflowRunStateQueued = { status: "queued", reason: fromState.reason };
 		stateTransitionEntries.push({
 			id: stateTransitionId,
 			workflowRunId: run.id,
 			type: "workflow_run",
 			status: "queued",
 			attempt: run.attempts,
-			state,
+			state: toState,
 		});
 		workflowRunUpdates.push({
 			filter: {
@@ -122,7 +139,6 @@ async function processChunk(
 				stateTransitionId,
 			},
 		});
-
 		outboxEntries.push({
 			id: ulid(),
 			namespaceId: run.namespaceId,
@@ -134,18 +150,13 @@ async function processChunk(
 		});
 	}
 
-	if (
-		!isNonEmptyArray(workflowRunIds) ||
-		!isNonEmptyArray(stateTransitionEntries) ||
-		!isNonEmptyArray(workflowRunUpdates)
-	) {
+	if (!isNonEmptyArray(stateTransitionEntries) || !isNonEmptyArray(workflowRunUpdates)) {
 		return;
 	}
 
 	const insertedOutboxEntries: WorkflowRunOutboxRowInsert[] = await repos.transaction(async (txRepos) => {
-		await txRepos.task.deleteStaleByWorkflowRunIds(workflowRunIds);
 		await txRepos.stateTransition.appendBatch(stateTransitionEntries);
-		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued("awaiting_retry", workflowRunUpdates);
+		const transitionedRunIds = await txRepos.workflowRun.bulkTransitionToQueued("scheduled", workflowRunUpdates);
 		const transitionedRunIdsSet = new Set(transitionedRunIds);
 		const outboxEntriesToInsert = outboxEntries.filter((entry) => transitionedRunIdsSet.has(entry.workflowRunId));
 		if (!isNonEmptyArray(outboxEntriesToInsert)) {
