+++ server/infra/db/pg/repository/workflow-run-outbox.ts
@@ -2,8 +2,8 @@ import type { NonEmptyArray } from "@aikirun/lib/array";
 import { isNonEmptyArray } from "@aikirun/lib/array";
 import type { WorkflowRunId } from "@aikirun/types/workflow-run";
 import { and, eq, inArray, lt, sql } from "drizzle-orm";
-import type { TimerStreamCursor } from "server/crons/lib/timer-stream";
-import type { CronContext } from "server/middleware/context";
+import type { TimerStreamCursor } from "server/daemons/lib/timer-stream";
+import type { DaemonContext } from "server/middleware/context";
 
 import { timerStreamCursorFilter } from "./lib/timer-stream";
 import type { PgDb } from "../provider";
@@ -19,7 +19,7 @@ export function createWorkflowRunOutboxRepository(db: PgDb) {
 		},
 
 		async listPending(
-			_context: CronContext,
+			_context: DaemonContext,
 			limit: number,
 			cursor?: TimerStreamCursor
 		): Promise<WorkflowRunOutboxRow[]> {
@@ -51,7 +51,7 @@ export function createWorkflowRunOutboxRepository(db: PgDb) {
 		},
 
 		async listStalePublished(
-			_context: CronContext,
+			_context: DaemonContext,
 			claimMinIdleTimeMs: number,
 			limit: number,
 			cursor?: TimerStreamCursor
